+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2024-07-30
+
+### Changed
+
+- Upgraded to Unity 2022.3.38f1, Burst 1.8.16, and URP 14.0.11.
+- Changed `HexCell` to a struct type.
+- Moved data and functionality from `HexCell` to `HexGrid`.
+
+### Removed
+
+- Intermediary cell properties for flags and values.
+
 ## [3.3.0] - 2024-04-09
 
 ### Added